feat(atlasDataDictionaryButton): moved butotn to AtlasDictionaryDirectory
