+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryContainer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AtlasDataDictionaryButton from '../AtlasDataDictionaryButton/AtlasDataDictionaryButton';
+import AtlasDataDictionaryButton from './AtlasDataDictionaryButton/AtlasDataDictionaryButton';
 
 const AtlasDataDictionaryContainer = () => <div style={{ width: '100%' }}><AtlasDataDictionaryButton /></div>;
